Add base stuff for a block with connected textures
Only adds "full", no horizontal or alike. All the models are done using a Forge blockstate for simplicity, so copying the example Blockstate and changing the textures is the easiest way to add a model

This is based off the tutorial by Darkhax linked below, though with a few things modified to make it easier for extending and support adding via EnumBlocks. http://tutorials.darkhax.net/connected-block-textures.html