Improve exceptions thrown on module lifecycle events (#105)
Make exceptions thrown by events in modules look nicer by removing the awkward InvocationTargetException wrapper. InvocationTargetException is a wrapper used when a method invoked by reflection itself throws an exception. This lets the log be a bit nicer as we don't have a useless part of the stacktrace just telling us an exception was thrown

Also rethrow runtime exceptions as is so we don't add the exception handler to the stack trace unnecessarily