Add fluid container ingredient
Matches a fluid container item stack that has the given fluid stack inside. Performs a strict match, if the item stack does not have exactly the right amount of fluid it will fail, or if the drained item stack is not the container item
Should for example match TiC copper cans or Ceramics clay buckets perfectly.