Attempt to fix edge cases in columns not lining up in generated indexes
Reported issue is SlimeKnights/TinkersConstruct#4720
Mantle text drawing logic for list elements includes a prefix element which is drawn before the text ignoring the split. In some other languages, the recommend split is after the prefix element causing the estimate to be too many lines, while the renderer never splits after the prefix.
To fix this issue, just ask Mantle directly for the number of lines needed to draw the text considering the prefix instead of using the Minecraft font renderer