Improve ItemOutput serialization and deserialization
Now takes full advantage of ItemStackLoadable to allow 2 loading options:
* Optional output is now handled directly in the API, so you can have an ItemOutput that resolves to empty, or disallow that entirely. Note tag output may still resolve to empty due to an empty tag.
* Fixed count is now an option, as some cases using ItemOutput don't support stack sizes greater than 1 (e.g. casting recipes).

In addition, ItemOutput.EMPTY now exists for those wanting an empty instance, and the default serialize function takes a boolean for whether to write count.
Left the old networking logic in place for now, it will be less compact as it always writes count and writes a boolean for empty (instead of just using items.AIR for empty). However, its easier to invoke so figured I'd leave it until more stuff migrates to loadables