Implement EitherLoadable
This is designed for things like FluidIngredient and EntityIngredient that have multiple forms controlled by subclasses. USing this let me much simplify the impl for both getting loadables
I did not use it for ItemOutput as that one wants 4 different loadable variants, which conflicts with how the either internals work. Only way to deal with this breaks the datagen only ingredients so I figured I'd stick with what works.