Rename FriendlyByteBuffer methods on Loadable to match StreamCodec
Saves me from doing a refactor later making it less effort for addons to update that directly call the methods.
As part of implementation, created Streamable, which is a less generic version of StreamCodec. I would just use StreamCodec but it has the minor issue of not existing.
EitherLoadable now uses streamable instead of loadable for its network logic as that makes it easier to implement a custom variant.
GenericLoaderRegistry and GenericNamedComponent both have their methods updated to the new name with the old ones kept around as deprecated.
This commit also marks the deprecation of IGenericLoader, come 1.20 I'll be making GenericLoaderRegistry a registry of RecordLoadable.