Implement full maps in loadables
To make this work, added a new sub-interface of Loadable: StringLoadable. Represents any loadable that can be written directly into a string or read from a string.
MapLoadables require a string loadable as a key, but accept any loadable as a value. Intentionally did not implement RecordLoadable for them as the map keys will conflict with other objects
Strings, resource locations, booleans, enums, colors, and tag keys all implement StringLoadable and work as keys
Registries and named component registry both implement StringLoadable via the helper ResourceLocationLoadable
IntLoadable does not implement StringLoadable, but has a new asString() method to create a loadable for strings as int, with a radix argument.
Since StringLoadable used to be a record for different max lengths, that logic was moved to a package private class, can be built using a static helper on StringLoadable